% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midisonifyR-package.R
\docType{package}
\name{midisonifyR-package}
\alias{midisonifyR}
\alias{midisonifyR-package}
\title{midisonifyR: Sonify dataframes with MIDI}
\description{
Sonify dataframes with MIDI somewhat similar fo ggplot. Map variables to pitch, volume, instruments and more.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/neuwirthe/midisonifyR}
  \item Report bugs at \url{https://github.com/neuwirthe/midisonifyR/issues}
}

}
\author{
\strong{Maintainer}: Erich Neuwirth \email{erich.neuwirth@univie.ac.at} [copyright holder]

}
\keyword{internal}
