% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/note_events.R
\name{sonify_variables}
\alias{sonify_variables}
\alias{note_events}
\alias{drum_events}
\title{Notes and drums from variables}
\usage{
note_events(
  data = NULL,
  mapping = props(),
  instrument = 1,
  base_time = 1,
  start_time = 0,
  base_pitch = 0,
  channel = 1,
  default_velocity = 0.8
)

drum_events(
  data,
  mapping = props(),
  drum = 35,
  base_time = 1,
  start_time = 0,
  default_velocity = 0.8,
  drum_bank = 0
)
}
\arguments{
\item{data}{dataframe to be mapped to notes}

\item{mapping}{list of mappings of variables to musical properties}

\item{instrument}{instrument number to be used for note events}

\item{base_time}{for time = 1 in milliseconds}

\item{start_time}{start time for playing notes}

\item{base_pitch}{offset for pitches}

\item{channel}{channel for MIDI notes.}

\item{default_velocity}{default velocity if not given through mapping}

\item{drum}{drum number from drum bank}

\item{drum_bank}{drum bank to be used}
}
\value{
midi_event dataframe
}
\description{
Map dataframe variables to musical notes
}
